fix product class filter
