+++ src/app/[locale]/list/_components/AasListComparisonHeader.tsx
@@ -24,7 +24,7 @@ export const AasListComparisonHeader = (props: CompareAasListBarType) => {
 
     return (
         <>
-            <Box display="flex" gap={2} alignItems="center">
+            <Box display="flex" gap={2} alignItems="center" justifyContent="flex-end">
                 {selectedAasList?.map((selectedAas) => (
                     <Box display="flex" flexDirection="row" alignItems="center" key={selectedAas}>
                         <Typography data-testid={`selected-${selectedAas}`}>{tooltipText(selectedAas, 15)}</Typography>
